+++ checker/src/main/java/org/checkerframework/checker/optional/OptionalImplAnnotatedTypeFactory.java
@@ -21,8 +21,8 @@ import org.checkerframework.framework.type.AnnotatedTypeMirror;
 import org.checkerframework.javacutil.AnnotationBuilder;
 import org.checkerframework.javacutil.TreeUtils;
 
-/** OptionalAnnotatedTypeFactory for the Optional Checker. */
-public class OptionalAnnotatedTypeFactory extends BaseAnnotatedTypeFactory {
+/** Type factory for the OptionalImplChecker. */
+public class OptionalImplAnnotatedTypeFactory extends BaseAnnotatedTypeFactory {
 
   /** The element for java.util.Optional.map(). */
   private final ExecutableElement optionalMap;
@@ -31,11 +31,12 @@ public class OptionalAnnotatedTypeFactory extends BaseAnnotatedTypeFactory {
   protected final AnnotationMirror PRESENT = AnnotationBuilder.fromClass(elements, Present.class);
 
   /**
-   * Creates an OptionalAnnotatedTypeFactory.
+   * Creates an OptionalImplAnnotatedTypeFactory.
    *
-   * @param checker the Optional Checker associated with this type factory
+   * @param checker the {@link org.checkerframework.checker.optional.OptionalImplChecker} associated
+   *     with this type factory
    */
-  public OptionalAnnotatedTypeFactory(BaseTypeChecker checker) {
+  public OptionalImplAnnotatedTypeFactory(BaseTypeChecker checker) {
     super(checker);
     postInit();
     optionalMap = TreeUtils.getMethodOrNull("java.util.Optional", "map", 1, getProcessingEnv());
@@ -49,13 +50,13 @@ public class OptionalAnnotatedTypeFactory extends BaseAnnotatedTypeFactory {
 
   /**
    * If {@code tree} is a call to {@link java.util.Optional#map(Function)} whose argument is a
-   * method reference, then this method adds {@code @Present} to {@code type} if the following is
+   * method reference m, then this method adds {@code @Present} to {@code type} if the following is
    * true:
    *
    * <ul>
    *   <li>The type of the receiver to {@link java.util.Optional#map(Function)} is {@code @Present},
    *       and
-   *   <li>{@link #returnHasNullable(MemberReferenceTree)} returns false.
+   *   <li>{@link #isReturnTypeNullable(MemberReferenceTree)} returns false for m.
    * </ul>
    *
    * @param tree a tree
@@ -67,20 +68,21 @@ public class OptionalAnnotatedTypeFactory extends BaseAnnotatedTypeFactory {
     }
     MethodInvocationTree mapTree = (MethodInvocationTree) tree;
     ExpressionTree argTree = mapTree.getArguments().get(0);
-    if (argTree.getKind() == Kind.MEMBER_REFERENCE) {
-      MemberReferenceTree memberReferenceTree = (MemberReferenceTree) argTree;
-      AnnotatedTypeMirror optType = getReceiverType(mapTree);
-      if (optType == null || !optType.hasEffectiveAnnotation(Present.class)) {
-        return;
-      }
-      if (!returnHasNullable(memberReferenceTree)) {
-        // The method still could have a @PolyNull on the return and might return null.
-        // If @PolyNull is the primary annotation on the parameter and not on any type
-        // arguments or array elements, then it is still safe to mark the optional type as
-        // present.
-        // TODO: Add the check for poly null on arguments.
-        type.replaceAnnotation(PRESENT);
-      }
+    if (argTree.getKind() != Kind.MEMBER_REFERENCE) {
+      return;
+    }
+    AnnotatedTypeMirror mapReceiver = getReceiverType(mapTree);
+    if (mapReceiver == null || !mapReceiver.hasEffectiveAnnotation(Present.class)) {
+      return;
+    }
+    MemberReferenceTree memberReferenceTree = (MemberReferenceTree) argTree;
+    if (!isReturnTypeNullable(memberReferenceTree)) {
+      // The method still could have a @PolyNull on the return and might return null.
+      // If @PolyNull is the primary annotation on the parameter and not on any type
+      // arguments or array elements, then it is still safe to mark the optional type as
+      // present.
+      // TODO: Add the check for @PolyNull on arguments.
+      type.replaceAnnotation(PRESENT);
     }
   }
 
@@ -88,18 +90,19 @@ public class OptionalAnnotatedTypeFactory extends BaseAnnotatedTypeFactory {
    * Returns true if the return type of the function type of {@code memberReferenceTree} is
    * annotated with {@code @Nullable}.
    *
-   * @param memberReferenceTree a member reference
+   * @param memberReferenceTree a member reference (which is to a method or constructor)
    * @return true if the return type of the function type of {@code memberReferenceTree} is
    *     annotated with {@code @Nullable}
    */
-  private boolean returnHasNullable(MemberReferenceTree memberReferenceTree) {
+  private boolean isReturnTypeNullable(MemberReferenceTree memberReferenceTree) {
+    // A member reference refers to either a method or a constructor.
     if (TreeUtils.MemberReferenceKind.getMemberReferenceKind(memberReferenceTree)
         .isConstructorReference()) {
       return false;
     }
     ExecutableElement memberReferenceFuncType = TreeUtils.elementFromUse(memberReferenceTree);
     if (memberReferenceFuncType.getEnclosingElement().getKind() == ElementKind.ANNOTATION_TYPE) {
-      // Annotation element accessor are always non-null;
+      // Annotation element accessors are always non-null;
       return false;
     }
 
@@ -111,10 +114,10 @@ public class OptionalAnnotatedTypeFactory extends BaseAnnotatedTypeFactory {
   }
 
   /**
-   * Returns true if {@code annos} contains a nullable annotation.
+   * Returns true if {@code annos} contains an annotation named "Nullable".
    *
    * @param annos a collection of annotations
-   * @return true if {@code annos} contains a nullable annotation
+   * @return true if {@code annos} contains an annotation named "Nullable".
    */
   private boolean containsNullable(Collection<? extends AnnotationMirror> annos) {
     for (AnnotationMirror anno : annos) {
@@ -128,6 +131,6 @@ public class OptionalAnnotatedTypeFactory extends BaseAnnotatedTypeFactory {
   @Override
   public CFTransfer createFlowTransferFunction(
       CFAbstractAnalysis<CFValue, CFStore, CFTransfer> analysis) {
-    return new OptionalTransfer(analysis);
+    return new OptionalImplTransfer(analysis);
   }
 }