+++ checker/tests/optional/Marks3a.java
@@ -1,5 +1,6 @@
 import java.util.List;
 import java.util.Optional;
+import org.checkerframework.dataflow.qual.*;
 
 /**
  * Test case for rule #3: "Prefer alternative APIs over Optional.isPresent() and Optional.get()."
@@ -11,13 +12,22 @@ public class Marks3a {
       return 42;
     }
 
+    @Pure
     String getName() {
+      return "Fozzy Bear";
+    }
+
+    String getNameImpure() {
+      System.out.println("Side effect");
       return "Fozzy Bear";
     }
   }
 
   String customerNameByID_acceptable(List<Customer> custList, int custID) {
     Optional<Customer> opt = custList.stream().filter(c -> c.getID() == custID).findFirst();
+
+    // Not valid to report a map.and.orelse warning here
+    String s = opt.isPresent() ? opt.get().getNameImpure() : "UNKNOWN";
 
     // :: warning: (prefer.map.and.orelse)
     return opt.isPresent() ? opt.get().getName() : "UNKNOWN";
@@ -28,6 +38,36 @@ public class Marks3a {
 
     // :: warning: (prefer.map.and.orelse)
     return !opt.isPresent() ? "UNKNOWN" : opt.get().getName();
+  }
+
+  String customerNameByID_acceptable3(List<Customer> custList, int custID) {
+    Optional<Customer> opt = custList.stream().filter(c -> c.getID() == custID).findFirst();
+
+    String customerName;
+    // :: warning: (prefer.map.and.orelse)
+    if (opt.isPresent()) {
+      customerName = opt.get().getName();
+    } else {
+      customerName = "UNKNOWN";
+    }
+
+    return customerName;
+  }
+
+  String customerNameByID_acceptable4(List<Customer> custList, int custID) {
+    Optional<Customer> opt = custList.stream().filter(c -> c.getID() == custID).findFirst();
+
+    String customerName = "";
+    String unknownCustomerName = "";
+
+    // This is OK, because the two LHSes differ.
+    if (opt.isPresent()) {
+      customerName = opt.get().getName();
+    } else {
+      unknownCustomerName = "UNKNOWN";
+    }
+
+    return customerName;
   }
 
   String customerNameByID_better(List<Customer> custList, int custID) {