+++ docs/CHANGELOG.md
@@ -3,6 +3,12 @@ Version 3.48.5 (February 3, 2025)
 
 **User-visible changes:**
 
+The Optional Checker is more precise for `Optional` values resulting from
+operations on container types (e.g., `List`, `Map`, `Iterable`).  It supports
+two new annotations:
+  * `@NonEmpty`
+  * `@UnknownNonEmpty`
+
 **Implementation details:**
 
 **Closed issues:**